requirements.txt
click==8.1.8
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
